Moving utilitary funcs to utils
